Some styling changes
