+++ extensions/meta/__init__.py
@@ -2,9 +2,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from .avatar import Avatar
 from .botinfo import BotInformation
-from .roleinfo import RoleInfo
 from .serverinfo import ServerInfo
 from .userinfo import Userinfo
 
@@ -12,7 +10,7 @@ if TYPE_CHECKING:
     from bot import Mafuyu
 
 
-class Meta(Avatar, BotInformation, RoleInfo, Userinfo, ServerInfo, name='Meta'):
+class Meta(BotInformation, Userinfo, ServerInfo, name='Meta'):
     """For everything related to Discord or Mafuyu."""
